Add 롬복

+++ calculator/src/main/java/repository/CalculatorRepository.java
@@ -3,8 +3,8 @@ package repository;
 import java.util.List;
 
 public interface CalculatorRepository {
-    void save(String exp, int result);
-    int getResult(String exp);
+    void save(String exp, double result);
+    double getResult(String exp);
     List<String> getResults();
     void clear();
 }
